Finition du CreateRestControlleur

+++ src/main/java/fr/litopia/cyberchamis/model/entity/Arret.java
@@ -59,19 +59,19 @@ public class Arret {
         this.ville = ville;
     }
 
-    public long getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(long latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public long getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(long longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
